requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
reportlab>=4.0.0
click>=8.1.0
python-dateutil>=2.8.0
spacy>=3.7.0
nltk>=3.8.0
lxml>=4.9.0